requirements.txt
asgiref==3.8.1
contourpy==1.2.1
cycler==0.12.1
Django==5.0.6
fonttools==4.53.1
google-auth==2.22.0
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==2.0.1
packaging==24.1
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
pytest==7.5.1
pytest-django==5.0.0
python-dotenv==1.0.0
pydantic==2.3.0
loguru==0.7.0
django-environ==1.0.0
